refactor: base classes
